Consolidate merkle tree scripts into one, add README, and introduce metadata.json
