+++ data/metadata.json
@@ -1,0 +1,80 @@
+{
+  "rootSigners": [
+    "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
+    "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
+    "0x3C44CdDdB6a900fa2b585dd299e03d12FA4293BC",
+    "0x90F79bf6EB2c4f870365E785982E1f101E93b906",
+    "0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65",
+    "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc",
+    "0x976EA74026E726554dB657fA54763abd0C3a0aa9",
+    "0x14dC79964da2C08b23698B3D3cc7Ca32193d9955",
+    "0x23618e81E3f5cdF7f54C3d65f7FBc0aBf5B21E8f",
+    "0xa0Ee7A142d267C1f36714E4a8F75612F20a79720"
+  ],
+  "merkleTrees": {
+    "priceMT": {
+      "timestamp": 1676940000,
+      "dataTypes": ["string", "uint256", "bytes32", "uint256", "uint256"],
+      "values": [
+        ["dapiName1", "1", "0x1234567890123456789012345678901234567890123456789012345678901234", "3600", "1000"],
+        ["dapiName2", "2", "0x2345678901234567890123456789012345678901234567890123456789012345", "7200", "2000"]
+      ],
+      "merkleRoot": "",
+      "signatures": {}
+    },
+    "dapiManagementMT": {
+      "timestamp": 1676940000,
+      "dataTypes": ["bytes32", "bytes32", "address"],
+      "values": [
+        [
+          "0x1234567890123456789012345678901234567890123456789012345678901234",
+          "0x1234567890123456789012345678901234567890123456789012345678901234",
+          "0x1DA48273Fc2ACd6f40b629538aE150Bea08cC7bB"
+        ],
+        [
+          "0x2345678901234567890123456789012345678901234567890123456789012345",
+          "0x2345678901234567890123456789012345678901234567890123456789012345",
+          "0x71B88fC1a8cce02B1339901CcC88C4234ee4cc57"
+        ]
+      ],
+      "merkleRoot": "",
+      "signatures": {}
+    },
+    "dapiFallbackMT": {
+      "timestamp": 1676940000,
+      "dataTypes": ["string", "bytes32", "address"],
+      "values": [
+        [
+          "dapiName1",
+          "0x1234567890123456789012345678901234567890123456789012345678901234",
+          "0x1DA48273Fc2ACd6f40b629538aE150Bea08cC7bB"
+        ],
+        [
+          "dapiName2",
+          "0x2345678901234567890123456789012345678901234567890123456789012345",
+          "0x71B88fC1a8cce02B1339901CcC88C4234ee4cc57"
+        ]
+      ],
+      "merkleRoot": "",
+      "signatures": {}
+    },
+    "apiIntegrationMT": {
+      "timestamp": 1676940000,
+      "dataTypes": ["address", "bytes32", "string"],
+      "values": [
+        [
+          "0x1DA48273Fc2ACd6f40b629538aE150Bea08cC7bB",
+          "0x1234567890123456789012345678901234567890123456789012345678901234",
+          "https://api.example.com/data"
+        ],
+        [
+          "0x71B88fC1a8cce02B1339901CcC88C4234ee4cc57",
+          "0x2345678901234567890123456789012345678901234567890123456789012345",
+          "https://api.example2.com/data"
+        ]
+      ],
+      "merkleRoot": "",
+      "signatures": {}
+    }
+  }
+}
