--- scripts/merkle/treeValues/dapiManagementMT.json
@@ -1,15 +1,0 @@
-{
-  "timestamp": 1676940000,
-  "values": [
-    [
-      "dapiName1",
-      "0x1234567890123456789012345678901234567890123456789012345678901234",
-      "0x1DA48273Fc2ACd6f40b629538aE150Bea08cC7bB"
-    ],
-    [
-      "dapiName2",
-      "0x2345678901234567890123456789012345678901234567890123456789012345",
-      "0x71B88fC1a8cce02B1339901CcC88C4234ee4cc57"
-    ]
-  ]
-}
