--- scripts/merkle/treeValues/priceMT.json
@@ -1,7 +1,0 @@
-{
-  "timestamp": 1676940000,
-  "values": [
-    ["dapiName1", "1", "0x1234567890123456789012345678901234567890123456789012345678901234", "3600", "1000"],
-    ["dapiName2", "2", "0x2345678901234567890123456789012345678901234567890123456789012345", "7200", "2000"]
-  ]
-}
